them comment + fix case test click button add nhieu lan

+++ 03.Front-end/src/app/features/user/adm003/adm003.component.ts
@@ -11,7 +11,6 @@ import { EmployeeService } from 'src/app/service/employee.service';
 import { BsModalService, BsModalRef } from 'ngx-bootstrap/modal';
 import { TemplateRef } from '@angular/core';
 import { MSG } from 'src/app/shared/utils/messages.constants';
-import { EmployeeFormControls } from 'src/app/shared/utils/form-control-names.constant';
 import { PAGE } from 'src/app/shared/utils/mode-constant';
 
 @Component({
@@ -30,11 +29,11 @@ import { PAGE } from 'src/app/shared/utils/mode-constant';
 export class ADM003Component {
   modalRef!: BsModalRef; // Tham chiếu đến modal Bootstrap đang được hiển thị
   MSG = MSG; // Hằng số chứa các message hiển thị
-
   listDepartments: Department[] = [];  // Danh sách các phòng ban, được dùng để hiển thị trong dropdown
   listCertifications: Certification[] = [];  // Danh sách các trình độ tiếng nhật, được dùng để hiển thị trong dropdown
   employeeData: any; // Dữ liệu thông tin của nhân viên
   employeeId!: number; // Id của nhân viên được chuyển từ ADM002 sang
+  isDeleted: boolean = false; // Biến để kiếm soát trạng thái khi người dùng click liên tục button [OK] để xóa nhân viên
 
   /**
    * Constructor khởi tạo component, inject các service cần thiết.
@@ -57,32 +56,29 @@ export class ADM003Component {
    * - Nếu không hợp lệ thì điều hướng sang trang SystemError.
    */
   ngOnInit(): void {
-    // Lấy id nhân viên truyền sang từ ADM002
-    this.employeeId = history.state?.employeeId;
-    if (isNaN(Number(this.employeeId)) || !this.employeeId) {
-      this.router.navigate(['error']);
+    this.employeeId = history.state?.employeeId;  // Lấy id nhân viên truyền sang từ ADM002
+    if (isNaN(Number(this.employeeId)) || !this.employeeId) { // Nếu id rỗng hoặc không phải số
+      this.router.navigate(['error']); // thì điều hướng đến màn hình System Error
     }
 
-    this.getEmployeeById(this.employeeId);
+    this.getEmployeeById(this.employeeId); // Gọi API để lấy thông tin chi tiết của nhân viên theo ID được truyền sang. 
   }
 
   /**
    * Gọi API backend để lấy thông tin chi tiết của nhân viên theo ID.
+   * Nếu thành công thì biến đổi dữ liệu và lấy dữ liệu nhân viên
+   * Nếu thất bại thì điều hướng đến màn hình System Error với mã lỗi được trả về từ response
    *
    * @param {number} employeeId - ID của nhân viên cần truy vấn.
    */
   getEmployeeById(employeeId: number) {
-    // Gọi service để lấy thông tin nhân viên theo ID.
-    this.employeeService.getEmployeeById(employeeId).subscribe({
-      next: (data) => {
-        // Xử lý dữ liệu để chỉ giữ chứng chỉ cao nhất.
-        this.transformData(data);
-        this.employeeData = data;
+    this.employeeService.getEmployeeById(employeeId).subscribe({  // Gọi service để lấy thông tin nhân viên theo ID.
+      next: (data) => { // Nếu thành công
+        this.transformData(data); // Xử lý dữ liệu để chỉ giữ chứng chỉ cao nhất.
+        this.employeeData = data; // Lấy dữ liệu chi tiết nhân viên sau khi đã biến đổi.
       },
-      error: (err) => {
-        console.log(err);
-        // Nếu xảy ra lỗi, điều hướng đến SystemError và truyền mã lỗi.
-        this.router.navigate(['error'], {
+      error: (err) => { // Nếu thất bại
+        this.router.navigate(['error'], { // Điều hướng đến System Error và truyền mã lỗi tương ứng từ response trả về.
           state: { errorCode: err?.error?.message?.code }
         });
       }
@@ -114,14 +110,19 @@ export class ADM003Component {
     this.router.navigate(['/user/list']);
   }
 
+  /**
+   * Điều hướng sang màn hình ADM004 và truyền giá trị cho employeeId và fromPage để màn ADM004 xử lý
+   */
   handleEdit() {
-    this.router.navigate(['/user/adm004'], {
-      state: { employeeId: this.employeeId, fromPage: PAGE.ADM003 }
+    this.router.navigate(['/user/adm004'], { // Điều hướng đến màn hình ADM004
+      state: { employeeId: this.employeeId, fromPage: PAGE.ADM003 } // employeeId là Id của nhân viên, fromPage để biết di chuyển từ màn hình nào sang ADM004
     });
   }
 
   /**
    * Xử lý hành động xóa nhân viên.
+   * Nếu thành công thì điều hướng đến màn hình ADM006
+   * Nếu thất bại thì điều hướng đến màn hình System Error
    * 
    * - Thêm class `closing` để tạo hiệu ứng đóng modal.
    * - Ẩn modal thông qua `BsModalRef`.
@@ -130,26 +131,29 @@ export class ADM003Component {
    * - Nếu thất bại, điều hướng đến màn SystemError với mã `errorCode`.
    */
   handleDelete() {
+    if (this.isDeleted) return; // Nếu đã xóa trước đó rồi thì không làm gì cả để trách gọi API xóa nhiều lần liên tục
+    this.isDeleted = true; // Cập nhật trạng thái bản ghi đã được thao tác xóa
+
     // Thêm hiệu ứng đóng trước
     const modal = document.querySelector('.modal-content');
     modal?.classList.add('closing');
-    this.modalRef?.hide();
+    this.modalRef?.hide(); // Ẩn modal
 
     setTimeout(() => {
       // Gọi API sau khi modal ẩn
       this.employeeService.deleteEmployeeById(this.employeeId).subscribe({
-        next: (data) => {
-          this.router.navigate(['user/adm006'], {
-            state: { completeCode: data?.message?.code }
+        next: (data) => { // Nếu thành công
+          this.router.navigate(['user/adm006'], { // Điều hướng đến màn ADM006
+            state: { completeCode: data?.message?.code } // Truyền giá trị mã thông báo được trả về từ response để hiển thị nội dung thông báo ở màn ADM006
           });
         },
-        error: (err) => {
-          this.router.navigate(['error'], {
-            state: { errorCode: err?.error?.message?.code }
+        error: (err) => { // Nếu thất bại
+          this.router.navigate(['error'], { // Điều hướng đến màn System Error 
+            state: { errorCode: err?.error?.message?.code } // Truyền giá trị mã lỗi được trả về từ response để hiển thị nội dung lỗi ở màn System Error 
           });
         }
       });
-    }, 400);
+    }, 400); // Gọi sau 400 ms để modal đóng
   }
 
   /**
